# docs/order-api-openapi.yaml
openapi: 3.1.0
info:
  title: Order Lifecycle API
  version: 1.0.0
  summary: Service-to-service REST API for managing the order lifecycle on top of Shopware 6.
  description: |
    A versioned, machine-to-machine REST API for the order lifecycle on top of
    Shopware 6. Phase 1 is implemented as a Shopware plugin that registers its
    own routes inside the Shopware backend (see README). Later phases evolve
    the deployment topology without breaking this contract.

    Not intended for human/browser use. All requests require TLS 1.2+.

    ### Authentication — current (Phase 1) vs. target

    - **Phase 1 (current):** Shopware OAuth 2.0 at `/api/oauth/token` (grants
      `password` and `client_credentials`). Bearer token on every request. Use
      the `shopwareOAuth2` security scheme.
    - **Target (Phase 5):** API gateway terminating **mTLS**, with **OAuth 2.0
      client credentials** from a dedicated authorisation server for fine-grained
      scopes, and **API keys** as a fallback. Use the `mTLS`,
      `oauth2ClientCredentials`, and `apiKey` schemes.

    Operations advertise both sets so the same contract describes both stages.
    Operators are expected to enable exactly one set at a time.

    ### Identifier format

    All `*Id` fields use Shopware's identifier convention: **32 lowercase
    hexadecimal characters, no dashes** (e.g. `019e79e17f17714883792eb3b8573ab6`).
    This is not an RFC 4122 UUID. The spec keeps `format: uuid` on field
...

...

...

...
  contact:
...
servers:
...
security:
  - shopwareOAuth2: []
  - mTLS: []
    oauth2ClientCredentials: [orders:read]
  - mTLS: []
    apiKey: []
paths:
  /orders:
    get:
      tags: [Orders]
      summary: List orders
      operationId: listOrders
      description: |
        Cursor-paginated list of orders. Supports a small allow-list of filters and sort keys.
      security:
        - mTLS: []
          oauth2ClientCredentials: [orders:read]
        - mTLS: []
          apiKey: []
      parameters:
        - $ref: '#/components/parameters/Traceparent'
        - name: status
          in: query
          description: Filter by domain status.
          schema:
            $ref: '#/components/schemas/OrderStatus'
        - name: customerId
          in: query
          schema:
            type: string
            format: uuid
        - name: salesChannelId
          in: query
          schema:
            type: string
            format: uuid
        - name: createdAfter
          in: query
          schema:
            type: string
            format: date-time
        - name: createdBefore
          in: query
          schema:
            type: string
            format: date-time
        - name: sort
          in: query
          description: 'Format: `field:asc|desc`. Allowed fields: `createdAt`, `updatedAt`, `orderNumber`.'
          schema:
            type: string
            pattern: '^(createdAt|updatedAt|orderNumber):(asc|desc)$'
            default: createdAt:desc
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        '200':
          description: A page of orders.
          headers:
            RateLimit-Limit: { $ref: '#/components/headers/RateLimitLimit' }
            RateLimit-Remaining: { $ref: '#/components/headers/RateLimitRemaining' }
            RateLimit-Reset: { $ref: '#/components/headers/RateLimitReset' }
            X-Request-Id: { $ref: '#/components/headers/XRequestId' }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderPage'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '5XX': { $ref: '#/components/responses/ServerError' }
    post:
      tags: [Orders]
      summary: Create an order
      operationId: createOrder
      description: |
        Create a new order. Idempotent via the `Idempotency-Key` header — required.
        Replaying the same key with the same body returns the original response;
        same key with a different body returns 409.
      security:
        - mTLS: []
          oauth2ClientCredentials: [orders:write]
        - mTLS: []
          apiKey: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/Traceparent'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreate'
      responses:
        '201':
          description: Order created (synchronous path).
          headers:
            Location:
              description: URL of the new order.
              schema: { type: string, format: uri }
            ETag:
              schema: { type: string }
            X-Request-Id: { $ref: '#/components/headers/XRequestId' }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '202':
          description: |
            Accepted for async processing (queued). Returned when the facade is in
            queue-mode or when the client opts in via `Prefer: respond-async`.
            Subscribe via webhook (`order.created`) or poll the job URL.
          headers:
            Location:
              description: Job status URL.
              schema: { type: string, format: uri }
            X-Request-Id: { $ref: '#/components/headers/XRequestId' }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsyncJob'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503':
          description: Backpressure — Shopware or the write queue is saturated.
          headers:
            Retry-After:
              schema: { type: integer }
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '5XX': { $ref: '#/components/responses/ServerError' }
  /orders/{orderId}:
    parameters:
      - $ref: '#/components/parameters/OrderId'
      - $ref: '#/components/parameters/Traceparent'
    get:
      tags: [Orders]
      summary: Get an order
      operationId: getOrder
      security:
        - mTLS: []
          oauth2ClientCredentials: [orders:read]
        - mTLS: []
          apiKey: []
      parameters:
        - name: If-None-Match
          in: header
          schema: { type: string }
        - name: If-Modified-Since
          in: header
          schema: { type: string }
      responses:
        '200':
          description: The order.
          headers:
            ETag: { schema: { type: string } }
            Last-Modified: { schema: { type: string } }
            X-Request-Id: { $ref: '#/components/headers/XRequestId' }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '304':
          description: Not modified.
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '5XX': { $ref: '#/components/responses/ServerError' }
    patch:
      tags: [Orders]
      summary: Update mutable attributes of an order
      operationId: patchOrder
      description: |
        Partial update of mutable fields: `customerComment`, `tags`, `customFields`, addresses.
        Status is not changed here — use `PUT /orders/{id}/status`.
        Requires `If-Match` (optimistic concurrency) and `Idempotency-Key`.
      security:
        - mTLS: []
          oauth2ClientCredentials: [orders:write]
        - mTLS: []
          apiKey: []
      parameters:
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: '#/components/schemas/OrderPatch'
      responses:
        '200':
          description: Updated order.
          headers:
            ETag: { schema: { type: string } }
            X-Request-Id: { $ref: '#/components/headers/XRequestId' }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '412': { $ref: '#/components/responses/PreconditionFailed' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '5XX': { $ref: '#/components/responses/ServerError' }
    delete:
      tags: [Orders]
      summary: Cancel or hard-delete an order
      operationId: deleteOrder
      description: |
        Soft delete by default: transitions order to `cancelled` and retains the record.
        Pass `?hard=true` to permanently delete — requires scope `orders:hard_delete`.
        Hard deletion is rejected by Shopware for orders with completed payments
        (surfaced as 409 Conflict).
      security:
        - mTLS: []
          oauth2ClientCredentials: [orders:write]
        - mTLS: []
          apiKey: []
      parameters:
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/IdempotencyKey'
        - name: hard
          in: query
          schema: { type: boolean, default: false }
        - name: reason
          in: query
          description: Free-text reason recorded in the audit log.
          schema: { type: string, maxLength: 500 }
      responses:
        '204':
          description: Deleted (soft or hard).
          headers:
            X-Request-Id: { $ref: '#/components/headers/XRequestId' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '412': { $ref: '#/components/responses/PreconditionFailed' }
        '5XX': { $ref: '#/components/responses/ServerError' }
  /orders/{orderId}/status:
    parameters:
      - $ref: '#/components/parameters/OrderId'
      - $ref: '#/components/parameters/Traceparent'
    put:
      tags: [Status]
      summary: Transition the order's domain status
      operationId: setOrderStatus
      description: |
        Drives the domain state machine (`open` → `in_progress` → `completed`/`cancelled`).
        Invalid transitions return 409. The facade maps to the corresponding Shopware
        `order.state` transition internally.
      security:
        - mTLS: []
          oauth2ClientCredentials: [orders:status]
        - mTLS: []
          apiKey: []
      parameters:
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatusTransition'
      responses:
        '200':
          description: Transition applied.
          headers:
            ETag: { schema: { type: string } }
            X-Request-Id: { $ref: '#/components/headers/XRequestId' }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '412': { $ref: '#/components/responses/PreconditionFailed' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '5XX': { $ref: '#/components/responses/ServerError' }
...
components:
  securitySchemes:
    shopwareOAuth2:
      type: oauth2
      description: |
        **Phase 1 — current.** Authentication delegated to Shopware's built-in
        OAuth 2.0 server at `/api/oauth/token`. Two grant types in production:

        - `password` — admin user (dev/admin use only).
        - `clientCredentials` — dedicated Shopware Integration user
          (`admin: false`, restricted scope). Recommended for service-to-service.

        Scopes are managed inside Shopware via the integration user's ACL roles,
        not by `scope` strings in the token request.
      flows:
        password:
          tokenUrl: /api/oauth/token
          scopes: {}
        clientCredentials:
          tokenUrl: /api/oauth/token
          scopes: {}
    mTLS:
      type: mutualTLS
      description: |
        **Phase 5 — target.** Client presents an X.509 certificate signed by
        the API's trusted CA. Subject CN/SAN is matched against an allow-list
        and becomes the principal. Enforced at the API gateway / service mesh.
    oauth2ClientCredentials:
      type: oauth2
      description: |
        **Phase 5 — target.** OAuth 2.0 Client Credentials grant against a
        dedicated authorisation server (not Shopware). Short-lived JWT access
        tokens (≤ 1h). Use in addition to mTLS for fine-grained scope-based
        authorization.
      flows:
        clientCredentials:
          tokenUrl: https://auth.example.com/oauth/token
          refreshUrl: https://auth.example.com/oauth/token
          scopes:
            orders:read: Read orders and events.
            orders:write: Create and modify orders (excluding hard delete).
            orders:status: Transition status, payment-status, delivery-status.
...
  parameters:
    OrderId:
      name: orderId
      in: path
      required: true
      description: Order UUID (the facade's own id, 1:1 with Shopware order id).
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Client-generated UUIDv4. The server caches `{key, body-hash, response}` for 24h.
        Replay with same body → same response. Replay with different body → 409.
      schema:
        type: string
        format: uuid
...
    Traceparent:
      name: traceparent
      in: header
      required: false
      description: W3C Trace Context.
      schema:
        type: string
  headers:
    RateLimitLimit:
      schema: { type: integer }
      description: Requests allowed in the current window.
    RateLimitRemaining:
      schema: { type: integer }
      description: Requests remaining in the current window.
    RateLimitReset:
      schema: { type: integer }
      description: Seconds until the rate limit window resets.
    XRequestId:
      schema: { type: string }
      description: Server-generated request id for log correlation.
  responses:
    BadRequest:
      description: Malformed request.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: Missing or invalid credentials (mTLS cert / token / API key).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: Authenticated but lacks the required scope.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Resource does not exist.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: Resource state conflicts with the requested operation (e.g., invalid status transition, idempotency-key reused with different body).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    UnprocessableEntity:
      description: Request was syntactically valid but semantically invalid (field validation errors).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    TooManyRequests:
      description: Rate limit exceeded.
      headers:
        Retry-After:
          schema: { type: integer }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    ServerError:
      description: Unexpected server error.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    OrderStatus:
      type: string
      enum: [open, in_progress, completed, cancelled]
      description: |
        Domain status, mapped internally to Shopware `order.state`:
        `open` → open, `in_progress` → in_progress, `completed` → completed,
        `cancelled` → cancelled.
...
    OrderCreate:
      type: object
      required: [salesChannelId, currency, customer, billingAddress, lineItems]
      properties:
        salesChannelId:
          type: string
          format: uuid
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
        customer:
          $ref: '#/components/schemas/Customer'
        billingAddress:
          $ref: '#/components/schemas/Address'
        shippingAddress:
          allOf:
            - $ref: '#/components/schemas/Address'
          description: Defaults to billing address if omitted.
          nullable: true
        lineItems:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/LineItem'
        paymentMethodId:
          type: string
          format: uuid
          nullable: true
        shippingMethodId:
          type: string
          format: uuid
          nullable: true
        customerComment:
          type: string
          nullable: true
          maxLength: 5000
        tags:
          type: array
          items: { type: string }
        customFields:
          type: object
          additionalProperties: true
          description: Free-form caller-defined metadata. Mapped 1:1 to Shopware customFields.
...
    Order:
      type: object
      required: [id, orderNumber, status, paymentStatus, deliveryStatus, currency, total, lineItems, billingAddress, customer, createdAt, updatedAt, version]
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        orderNumber:
          type: string
          readOnly: true
          description: Human-facing order number, generated by Shopware.
        version:
          type: integer
          readOnly: true
          description: Monotonically increasing version, also returned as ETag.
        salesChannelId: { type: string, format: uuid, readOnly: true }
        status: { $ref: '#/components/schemas/OrderStatus' }
        paymentStatus: { $ref: '#/components/schemas/PaymentStatus' }
        deliveryStatus: { $ref: '#/components/schemas/DeliveryStatus' }
        currency: { type: string, pattern: '^[A-Z]{3}$' }
        subtotal: { $ref: '#/components/schemas/Money' }
        shipping: { $ref: '#/components/schemas/Money' }
        tax: { $ref: '#/components/schemas/Money' }
        total: { $ref: '#/components/schemas/Money' }
        customer: { $ref: '#/components/schemas/Customer' }
        billingAddress: { $ref: '#/components/schemas/Address' }
        shippingAddress:
          allOf:
            - $ref: '#/components/schemas/Address'
          nullable: true
        lineItems:
          type: array
          items: { $ref: '#/components/schemas/LineItem' }
        deliveries:
          type: array
          description: |
            Embedded shipment list for read convenience. Populated from the read
            projection — for orders with many split shipments, prefer the dedicated
            `/orders/{id}/deliveries` endpoint to avoid bloated payloads.
          items: { $ref: '#/components/schemas/Delivery' }
        customerComment: { type: string, nullable: true }
        tags:
          type: array
          items: { type: string }
        customFields:
          type: object
          additionalProperties: true
        createdAt: { type: string, format: date-time, readOnly: true }
        updatedAt: { type: string, format: date-time, readOnly: true }
    OrderPage:
      type: object
      required: [items, page]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Order' }
        page:
          type: object
          required: [limit]
          properties:
            limit: { type: integer }
            nextCursor:
              type: string
              nullable: true
              description: Cursor to fetch the next page; null when no more pages.
...
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          description: URI identifying the problem type; SHOULD resolve to documentation.
          example: https://errors.example.com/order/invalid_state_transition
          default: about:blank
        title:
          type: string
          description: Short, human-readable summary of the problem type.
        status:
          type: integer
          description: HTTP status code.
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence.
        instance:
          type: string
          format: uri
          description: URI identifying this specific occurrence.
        code:
          type: string
          description: Stable machine-readable error code.
          example: order.invalid_state_transition
        traceId:
          type: string
          description: Correlation id for log lookup.
        errors:
          type: array
          description: Field-level validation errors (used with 422).
          items:
            type: object
            required: [pointer, code, message]
            properties:
              pointer:
                type: string
                description: JSON Pointer (RFC 6901) to the offending field.
                example: /lineItems/0/quantity
              code:
                type: string
                example: must_be_positive
              message:
                type: string
                example: quantity must be >= 1
    AsyncJob:
      type: object
      description: Returned with 202 Accepted for async-processed mutations.
      required: [jobId, status, createdAt]
      properties:
        jobId:
          type: string
          format: uuid
        status:
          type: string
          enum: [queued, in_progress, succeeded, failed]
        resourceUrl:
          type: string
          format: uri
          nullable: true
          description: Set once the resource exists.
        createdAt: { type: string, format: date-time }
        error:
          allOf:
            - $ref: '#/components/schemas/Problem'
          nullable: true
...