# src/Plugin/OrderIntegration/Controller/OrderController.php
<?php declare(strict_types=1);

namespace Scotty42\OrderIntegration\Controller;

use Scotty42\OrderIntegration\Exception\OrderNotFoundException;
use Scotty42\OrderIntegration\Cqrs\CqrsGateway;
use Scotty42\OrderIntegration\Exception\ValidationException;
use Scotty42\OrderIntegration\Service\OrderCreationService;
use Scotty42\OrderIntegration\Service\OrderMapper;
use Scotty42\OrderIntegration\Service\OrderPatchService;
use Scotty42\OrderIntegration\Service\StateMachineService;
use Scotty42\OrderIntegration\Validator\QueryValidator;
use Shopware\Core\Checkout\Order\OrderEntity;
use Shopware\Core\Framework\Context;
use Shopware\Core\Framework\DataAbstractionLayer\EntityRepository;
use Shopware\Core\Framework\DataAbstractionLayer\Search\Criteria;
use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\EqualsFilter;
use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\MultiFilter;
use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\NotFilter;
use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\RangeFilter;
use Shopware\Core\Framework\DataAbstractionLayer\Search\Sorting\FieldSorting;
use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
use Symfony\Component\HttpFoundation\JsonResponse;
use Symfony\Component\HttpFoundation\Request;
use Symfony\Component\HttpFoundation\Response;
use Symfony\Component\Routing\Attribute\Route;

#[Route(defaults: ['_routeScope' => ['api']])]
class OrderController extends AbstractController
{
    public function __construct(
        private readonly EntityRepository $orderRepository,
        private readonly QueryValidator $queryValidator,
        private readonly OrderCreationService $orderCreationService,
        private readonly OrderPatchService $orderPatchService,
        private readonly OrderMapper $orderMapper,
        private readonly StateMachineService $stateMachineService,
        private readonly CqrsGateway $cqrs,
    ) {}

    #[Route(
        path: '/api/order-integration/v1/orders',
        name: 'api.order-integration.orders.list',
        methods: ['GET']
    )]
    public function list(Request $request, Context $context): JsonResponse
    {
        $this->queryValidator->validateListParams($request->query->all());

        $limit = min($request->query->getInt('limit', 50), 200);

        // CQRS read path: serve from the projection when enabled.
        if ($this->cqrs->projectionReadsEnabled()) {
            $filters = array_filter([
                'status'         => $request->query->get('status'),
                'salesChannelId' => $request->query->get('salesChannelId'),
            ], static fn ($v) => $v !== null);
            $projected = $this->cqrs->listProjected($filters, $limit, $request->query->get('cursor'));

            return new JsonResponse([
                'items' => $projected['items'],
                'page'  => ['limit' => $limit, 'nextCursor' => $projected['nextCursor']],
            ]);
        }

        $criteria = new Criteria();
        $criteria->addSorting(new FieldSorting('createdAt', FieldSorting::DESCENDING));
        $criteria->addSorting(new FieldSorting('id', FieldSorting::DESCENDING));
        $criteria->addAssociations(OrderMapper::REQUIRED_ASSOCIATIONS);
        $criteria->setLimit($limit + 1);

        if ($status = $request->query->get('status')) {
            $criteria->addFilter(new EqualsFilter('stateMachineState.technicalName', $status));
        }
        if ($customerId = $request->query->get('customerId')) {
            $criteria->addFilter(new EqualsFilter('orderCustomer.customerId', $customerId));
        }
        if ($createdAfter = $request->query->get('createdAfter')) {
            $criteria->addFilter(new RangeFilter('createdAt', [RangeFilter::GTE => $createdAfter]));
        }
        if ($createdBefore = $request->query->get('createdBefore')) {
            $criteria->addFilter(new RangeFilter('createdAt', [RangeFilter::LTE => $createdBefore]));
        }

        if ($cursorRaw = $request->query->get('cursor')) {
            $cursor = json_decode(base64_decode($cursorRaw), true);
            $criteria->addFilter(new MultiFilter(MultiFilter::CONNECTION_OR, [
                new RangeFilter('createdAt', [RangeFilter::LT => $cursor['createdAt']]),
                new MultiFilter(MultiFilter::CONNECTION_AND, [
                    new EqualsFilter('createdAt', $cursor['createdAt']),
                    new NotFilter(MultiFilter::CONNECTION_AND, [
                        new EqualsFilter('id', $cursor['id']),
                    ]),
                    new RangeFilter('id', [RangeFilter::LT => $cursor['id']]),
                ]),
            ]));
        }

        $orders = $this->orderRepository->search($criteria, $context);
        $elements = array_values($orders->getElements());

        $hasMore = count($elements) > $limit;
        if ($hasMore) {
            array_pop($elements);
        }

        $nextCursor = null;
        if ($hasMore && !empty($elements)) {
            $last = end($elements);
            $nextCursor = base64_encode(json_encode([
                'createdAt' => $last->getCreatedAt()?->format(\DateTimeInterface::ATOM),
                'id'        => $last->getId(),
            ]));
        }

        return new JsonResponse([
            'items' => array_map(fn(OrderEntity $order) => $this->orderMapper->mapOrder($order), $elements),
            'page'  => [
                'limit'      => $limit,
                'nextCursor' => $nextCursor,
            ],
        ]);
    }

    #[Route(
        path: '/api/order-integration/v1/orders',
        name: 'api.order-integration.orders.create',
        methods: ['POST']
    )]
    public function create(Request $request, Context $context): JsonResponse
    {
        $data = json_decode($request->getContent(), true);

        if (!is_array($data)) {
            throw new ValidationException([
                ['pointer' => '/', 'code' => 'invalid_json', 'message' => 'Request body must be a JSON object.'],
            ]);
        }

        $errors = [];
        if (empty($data['salesChannelId'])) {
            $errors[] = ['pointer' => '/salesChannelId', 'code' => 'required', 'message' => 'salesChannelId is required'];
        }
        if (empty($data['lineItems'])) {
            $errors[] = ['pointer' => '/lineItems', 'code' => 'required', 'message' => 'lineItems must not be empty'];
        }
        if (!empty($errors)) {
            throw new ValidationException($errors);
        }

        // CQRS write path: enqueue and return 202 when async writes are enabled
        // (or the caller sends Prefer: respond-async). Protects Shopware from
        // raw parallel write load — see docs/cqrs-write-queue-concept.md.
        if ($this->cqrs->wantsAsyncWrite($request)) {
            if ($this->cqrs->shouldShed()) {
                return new JsonResponse([
                    'type' => 'about:blank', 'title' => 'Service Unavailable', 'status' => 503,
                    'detail' => 'Write queue saturated, retry later.', 'code' => 'order.backpressure',
                ], Response::HTTP_SERVICE_UNAVAILABLE, [
                    'Content-Type' => 'application/problem+json',
                    'Retry-After'  => (string) $this->cqrs->retryAfterSeconds(),
                ]);
            }

            $job = $this->cqrs->enqueueOrderCreate($data, $request->headers->get('Idempotency-Key'));

            return new JsonResponse(
                ['jobId' => $job->id, 'status' => $job->status],
                Response::HTTP_ACCEPTED,
                ['Location' => sprintf('/api/order-integration/v1/jobs/%s', $job->id)]
            );
        }

        $orderId = $this->orderCreationService->createOrder($data, $context);
        $order = $this->findOrder($orderId, $context);

        return new JsonResponse(
            $this->orderMapper->mapOrder($order),
            Response::HTTP_CREATED,
            [
                'Location' => sprintf('/api/order-integration/v1/orders/%s', $orderId),
                'ETag'     => $this->orderMapper->etagFor($order),
            ]
        );
    }

    #[Route(
        path: '/api/order-integration/v1/orders/{orderId}',
        name: 'api.order-integration.orders.get',
        methods: ['GET']
    )]
    public function get(string $orderId, Context $context): JsonResponse
    {
        if ($this->cqrs->projectionReadsEnabled()) {
            $projected = $this->cqrs->getProjectedOrder($orderId);
            if ($projected !== null) {
                return new JsonResponse($projected, Response::HTTP_OK, [
                    'ETag' => $this->weakEtagFromSnapshot($projected),
                ]);
            }
        }

        $order = $this->findOrder($orderId, $context);

        return new JsonResponse(
            $this->orderMapper->mapOrder($order),
            Response::HTTP_OK,
            ['ETag' => $this->orderMapper->etagFor($order)]
        );
    }

    #[Route(
        path: '/api/order-integration/v1/orders/{orderId}',
        name: 'api.order-integration.orders.patch',
        methods: ['PATCH']
    )]
    public function patch(string $orderId, Request $request, Context $context): JsonResponse
    {
        $this->findOrder($orderId, $context); // assert exists

        $data = json_decode($request->getContent(), true) ?? [];

        $allowed = ['customerComment', 'customFields', 'tags', 'billingAddress', 'shippingAddress'];
        $data = array_intersect_key($data, array_flip($allowed));

        if (empty($data)) {
            throw new ValidationException([
                ['pointer' => '/', 'code' => 'no_mutable_fields', 'message' => 'No mutable fields provided.'],
            ]);
        }

        $this->orderPatchService->patch($orderId, $data, $context);

        $order = $this->findOrder($orderId, $context);

        return new JsonResponse(
            $this->orderMapper->mapOrder($order),
            Response::HTTP_OK,
            ['ETag' => $this->orderMapper->etagFor($order)]
        );
    }

    #[Route(
        path: '/api/order-integration/v1/orders/{orderId}',
        name: 'api.order-integration.orders.delete',
        methods: ['DELETE']
    )]
    public function delete(string $orderId, Request $request, Context $context): JsonResponse
    {
        $this->findOrder($orderId, $context);

        $hard = $request->query->getBoolean('hard', false);

        if ($hard) {
            return new JsonResponse([
                'type'   => 'about:blank',
                'title'  => 'Forbidden',
                'status' => 403,
                'detail' => 'Hard delete requires scope orders:hard_delete. Not yet implemented.',
                'code'   => 'order.hard_delete_not_permitted',
            ], Response::HTTP_FORBIDDEN, ['Content-Type' => 'application/problem+json']);
        }

        // Soft delete: transition to cancelled. If already cancelled, the
        // state-machine will reject the transition — treat as idempotent.
        try {
            $this->stateMachineService->transitionOrder($orderId, 'cancelled', $context);
        } catch (\Scotty42\OrderIntegration\Exception\InvalidTransitionException $e) {
            // Already cancelled — idempotent success.
        }

        return new JsonResponse(null, Response::HTTP_NO_CONTENT);
    }

    /**
     * Weak ETag for a projection snapshot (no entity available on this path).
     *
     * @param array<string,mixed> $snapshot
     */
    private function weakEtagFromSnapshot(array $snapshot): string
    {
        $material = ($snapshot['id'] ?? '') . '|' . ($snapshot['version'] ?? '') . '|' . ($snapshot['updatedAt'] ?? '');

        return 'W/"' . sha1($material) . '"';
    }

    private function findOrder(string $orderId, Context $context): OrderEntity
    {
        $criteria = new Criteria([$orderId]);
        $criteria->addAssociations(OrderMapper::REQUIRED_ASSOCIATIONS);

        $order = $this->orderRepository->search($criteria, $context)->first();

        if (!$order instanceof OrderEntity) {
            throw new OrderNotFoundException($orderId);
        }

        return $order;
    }
}